Migrate AccuWeather to new entity naming style (#75127)
* Use new entity naming style

* Move AccuWeatherSensorDescription and SENSOR consts to sensor platform

* Remove duplicate code

* Suggested change

* Format url
